refactor(web): move back/forward buttons into NavRail, remove NavigationBar
