+++ apps/web/e2e/navigation-history.spec.ts
@@ -2,7 +2,7 @@
  * @area navigation
  * @priority P1
  *
- * Tests for navigation history: back/forward buttons, breadcrumb, and navigation tracking.
+ * Tests for navigation history: back/forward buttons and navigation tracking.
  *
  * Prerequisites: pnpm dev running on localhost:5173, at least one vault exists.
  */
@@ -10,9 +10,9 @@
 import { test, expect } from '@playwright/test';
 
 test.describe('Navigation History', () => {
-  test('back and forward buttons are rendered in the top bar', async ({ page }) => {
+  test('back and forward buttons are rendered in the nav rail', async ({ page }) => {
     await page.goto('http://localhost:5173/');
-    await page.waitForSelector('[data-testid="nav-topbar"]');
+    await page.waitForSelector('[data-testid="nav-back"]');
 
     const backBtn = page.locator('[data-testid="nav-back"]');
     const forwardBtn = page.locator('[data-testid="nav-forward"]');
@@ -23,7 +23,7 @@ test.describe('Navigation History', () => {
 
   test('back button is disabled on initial page load', async ({ page }) => {
     await page.goto('http://localhost:5173/');
-    await page.waitForSelector('[data-testid="nav-topbar"]');
+    await page.waitForSelector('[data-testid="nav-back"]');
 
     const backBtn = page.locator('[data-testid="nav-back"]');
     const forwardBtn = page.locator('[data-testid="nav-forward"]');
@@ -34,7 +34,7 @@ test.describe('Navigation History', () => {
 
   test('back button becomes enabled after navigating to another page', async ({ page }) => {
     await page.goto('http://localhost:5173/');
-    await page.waitForSelector('[data-testid="nav-topbar"]');
+    await page.waitForSelector('[data-testid="nav-back"]');
 
     // Navigate to knowledge base
     await page.locator('[data-view="knowledge"]').click();
@@ -50,7 +50,7 @@ test.describe('Navigation History', () => {
 
   test('back button navigates to previous route', async ({ page }) => {
     await page.goto('http://localhost:5173/');
-    await page.waitForSelector('[data-testid="nav-topbar"]');
+    await page.waitForSelector('[data-testid="nav-back"]');
 
     // Navigate to history page
     await page.locator('[data-view="history"]').click();
@@ -75,18 +75,5 @@ test.describe('Navigation History', () => {
     // Click forward — should go to history
     await forwardBtn.click();
     await page.waitForURL('**/history');
-  });
-
-  test('breadcrumb shows current page name', async ({ page }) => {
-    await page.goto('http://localhost:5173/');
-    await page.waitForSelector('[data-testid="nav-topbar"]');
-
-    const breadcrumb = page.locator('.nav-topbar__breadcrumb');
-    await expect(breadcrumb).toHaveText('首页');
-
-    // Navigate to settings
-    await page.locator('[data-view="settings"]').click();
-    await page.waitForURL('**/settings');
-    await expect(breadcrumb).toHaveText('设置');
   });
 });
